Add posts to magazine and donation link

+++ wp-content/plugins/duplicate-post/duplicate-post-admin.php
@@ -12,7 +12,7 @@ require_once (dirname(__FILE__).'/compat/duplicate-post-jetpack.php');
  * Wrapper for the option 'duplicate_post_version'
 */
 function duplicate_post_get_installed_version() {
-	return get_site_option( 'duplicate_post_version' );
+	return get_option( 'duplicate_post_version' );
 }
 
 /**
@@ -89,7 +89,7 @@ function duplicate_post_admin_init(){
 function duplicate_post_plugin_upgrade() {
 	$installed_version = duplicate_post_get_installed_version();
 	
-	if ( $installed_version==duplicate_post_get_current_version() )
+	if ( $installed_version == duplicate_post_get_current_version() )
 		return;
 
 		
@@ -131,7 +131,7 @@ function duplicate_post_plugin_upgrade() {
 	add_option('duplicate_post_copytitle','1');
 	add_option('duplicate_post_copydate','0');
 	add_option('duplicate_post_copystatus','0');
-	add_option('duplicate_post_copyslug','1');
+	add_option('duplicate_post_copyslug','0');
 	add_option('duplicate_post_copyexcerpt','1');
 	add_option('duplicate_post_copycontent','1');
 	add_option('duplicate_post_copythumbnail','1');
@@ -177,8 +177,8 @@ function duplicate_post_plugin_upgrade() {
 	delete_option('duplicate_post_view_user_level');
 	delete_option('dp_notice');
 	
-	delete_option('duplicate_post_version');
-	update_site_option( 'duplicate_post_version', duplicate_post_get_current_version() );
+	delete_site_option('duplicate_post_version');
+	update_option( 'duplicate_post_version', duplicate_post_get_current_version() );
 	
 	delete_option('duplicate_post_show_notice', 0);
 	update_site_option('duplicate_post_show_notice', 1);
@@ -191,9 +191,10 @@ function duplicate_post_plugin_upgrade() {
 function duplicate_post_show_update_notice() {
 	if(!current_user_can( 'manage_options')) return;
 	$class = 'notice is-dismissible';
-	$message = '<strong><a href="https://duplicate-post.lopo.it/">'.esc_html__('Check out the new documentation for Duplicate Post!', 'duplicate-post').'</a></strong><br/>';
-	$message .= '<em>'.esc_html__('Duplicate Post is now also compatible with WPML!', 'duplicate-post').'</em><br/>';
-	$message .= '<strong>'.sprintf(wp_kses(__('Help me develop the plugin and provide support by <a href="%s">donating even a small sum</a>.', 'duplicate-post'), array( 'a' => array( 'href' => array() ) ) ), "https://duplicate-post.lopo.it/donate").'</strong>';
+	$message = '<strong>'.sprintf(__("What's new in Duplicate Post version %s:", 'duplicate-post'), DUPLICATE_POST_CURRENT_VERSION).'</strong><br/>';
+	$message .= esc_html__('Simple compatibility with Gutenberg user interface: enable "Admin bar" under the Settings', 'duplicate-post').' — '.esc_html__('"Slug" option unset by default on new installations', 'duplicate-post').'<br/>';
+	$message .= '<em><a href="https://duplicate-post.lopo.it/">'.esc_html__('Check out the documentation', 'duplicate-post').'</a> — '.sprintf(__('Please <a href="%s">review the settings</a> to make sure it works as you expect.', 'duplicate-post'), admin_url('options-general.php?page=duplicatepost')).'</em><br/>';
+	$message .= esc_html__('Serving the WordPress community since November 2007.', 'duplicate-post').' <strong>'.sprintf(wp_kses(__('Help me develop the plugin and provide support by <a href="%s">donating even a small sum</a>.', 'duplicate-post'), array( 'a' => array( 'href' => array() ) ) ), "https://duplicate-post.lopo.it/donate").'</strong>';
 	global $wp_version;
 	if( version_compare($wp_version, '4.2') < 0 ){
 		$message .= ' | <a id="duplicate-post-dismiss-notice" href="javascript:duplicate_post_dismiss_notice();">'.__('Dismiss this notice.').'</a>';
@@ -276,6 +277,10 @@ function duplicate_post_add_removable_query_arg( $removable_query_args ){
 * then redirects to the post list
 */
 function duplicate_post_save_as_new_post($status = ''){
+	if(!duplicate_post_is_current_user_allowed_to_copy()){
+		wp_die(esc_html__('Current user is not allowed to copy posts.', 'duplicate-post'));
+	}
+	
 	if (! ( isset( $_GET['post']) || isset( $_POST['post'])  || ( isset($_REQUEST['action']) && 'duplicate_post_save_as_new_post' == $_REQUEST['action'] ) ) ) {
 		wp_die(esc_html__('No post to duplicate has been supplied!', 'duplicate-post'));
 	}
@@ -592,6 +597,11 @@ function duplicate_post_create_duplicate($post, $status = '', $parent_id = '') {
 	if(!empty($increase_menu_order_by) && is_numeric($increase_menu_order_by)){
 		$menu_order += intval($increase_menu_order_by);
 	}
+	
+	$post_name = $post->post_name;
+	if(get_option('duplicate_post_copyslug') != 1){
+		$post_name = '';
+	}
 
 	$new_post = array(
 	'menu_order' => $menu_order,
@@ -607,6 +617,7 @@ function duplicate_post_create_duplicate($post, $status = '', $parent_id = '') {
 	'post_status' => $new_post_status,
 	'post_title' => $title,
 	'post_type' => $post->post_type,
+	'post_name' => $post_name
 	);
 
 	if(get_option('duplicate_post_copydate') == 1){
@@ -616,33 +627,22 @@ function duplicate_post_create_duplicate($post, $status = '', $parent_id = '') {
 
 	$new_post_id = wp_insert_post(wp_slash($new_post));
 
-	// If the copy is published or scheduled, we have to set a proper slug.
-	if ($new_post_status == 'publish' || $new_post_status == 'future'){
-		$post_name = $post->post_name;
-		if(get_option('duplicate_post_copyslug') != 1){
-			$post_name = '';
-		}
-		$post_name = wp_unique_post_slug($post_name, $new_post_id, $new_post_status, $post->post_type, $new_post_parent);
-
-		$new_post = array();
-		$new_post['ID'] = $new_post_id;
-		$new_post['post_name'] = $post_name;
-
-		// Update the post into the database
-		wp_update_post( wp_slash($new_post) );
-	}
-
 	// If you have written a plugin which uses non-WP database tables to save
 	// information about a post you can hook this action to dupe that data.
-	if ($post->post_type == 'page' || is_post_type_hierarchical( $post->post_type ))
-		do_action( 'dp_duplicate_page', $new_post_id, $post, $status );
-	else
-		do_action( 'dp_duplicate_post', $new_post_id, $post, $status );
-
-	delete_post_meta($new_post_id, '_dp_original');
-	add_post_meta($new_post_id, '_dp_original', $post->ID);
-
-	do_action('duplicate_post_post_copy');
+	
+	if($new_post_id !== 0 && !is_wp_error($new_post_id)){
+		
+		if ($post->post_type == 'page' || is_post_type_hierarchical( $post->post_type ))
+			do_action( 'dp_duplicate_page', $new_post_id, $post, $status );
+		else
+			do_action( 'dp_duplicate_post', $new_post_id, $post, $status );
+	
+		delete_post_meta($new_post_id, '_dp_original');
+		add_post_meta($new_post_id, '_dp_original', $post->ID);
+	
+		do_action('duplicate_post_post_copy');
+		
+	}
 	
 	return $new_post_id;
 }
